got buy and sell screen to work

+++ project/portfolio/models.py
@@ -29,7 +29,7 @@ class Holding( models.Model ):
     symbol = models.CharField( max_length=50 )
     date = models.DateField()
     price = models.FloatField()
-    shares = models.FloatField()
+    shares = models.IntegerField()
     portfolio = models.ForeignKey( Portfolio )
 
 class Snippet( models.Model ):
